attempt to fix issue with installing this gem via bundler
in case the rack gem is not on the machine already, and we are installing for
prod, it would fail. not sure if this works, but we should always have rack
when using the actual gem. the utils included didn't seem to be used, but i may
be wrong. in that case we can bring that back and put a add_runtime_dependency
into the gemspec